use default export, unpack values as needed, fixes #71
