+++ src/fixtures/constants.js
@@ -1,5 +1,15 @@
 const MEAL_PERIOD_DAYS = 37
 
-export default {
-  MEAL_PERIOD_DAYS,
+const RATINGS = {
+  "secure": 4,
+  "moderately insecure": 3,
+  "highly insecure": 2,
+  "extremely insecure": 1
 }
+
+const exports = {
+  MEAL_PERIOD_DAYS,
+  RATINGS
+}
+
+export default exports
